Converted to jetpack nav, single activity structure
