+++ app/src/main/AndroidManifest.xml
@@ -12,9 +12,6 @@
         android:theme="@style/Theme.Pinpoint"
         tools:targetApi="31">
         <activity
-            android:name=".ProfilePageActivity"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
