--- app/src/main/java/edu/wisc/ece/pinpoint/ProfilePageActivity.java
@@ -1,90 +1,0 @@
-package edu.wisc.ece.pinpoint;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.viewpager2.widget.ViewPager2;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
-
-import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.navigation.NavigationBarView;
-import com.google.android.material.tabs.TabLayout;
-
-public class ProfilePageActivity extends AppCompatActivity {
-
-    private TabLayout tabLayout;
-    private ViewPager2 viewPager;
-    private BottomNavigationView navBar;
-    private FloatingActionButton mapButton;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile_page);
-        navBar = findViewById(R.id.navBar);
-        navBar.getMenu().getItem(2).setEnabled(false);
-        navBar.setSelectedItemId(R.id.navbarProfile);
-        mapButton = findViewById(R.id.mapButton);
-        navBar.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                switch(item.getItemId()){
-                    case R.id.navbarProfile:
-                        return true;
-                    case R.id.navbarSearch:
-                        return true;
-                    case R.id.navbarLeaderboard:
-                        return true;
-                    case R.id.navbarFeed:
-                        return true;
-                }
-                return false;
-            }
-        });
-        mapButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(new Intent(getApplicationContext(), MainActivity.class));
-                overridePendingTransition(0,0);
-            }
-        });
-
-        tabLayout = findViewById(R.id.tab_layout);
-        viewPager = findViewById(R.id.view_pager);
-        tabLayout.addTab(tabLayout.newTab().setText("Activity"));
-        tabLayout.addTab(tabLayout.newTab().setText("Dropped Pins"));
-        ProfileFragmentAdapter fragmentAdapter = new ProfileFragmentAdapter(this.getSupportFragmentManager(),  tabLayout.getTabCount(), getLifecycle());
-        viewPager.setAdapter(fragmentAdapter);
-
-        tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                viewPager.setCurrentItem(tab.getPosition());
-            }
-
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-
-            }
-
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-
-            }
-        });
-
-        viewPager.registerOnPageChangeCallback(new ViewPager2.OnPageChangeCallback() {
-            @Override
-            public void onPageSelected(int position) {
-                super.onPageSelected(position);
-                tabLayout.getTabAt(position).select();
-            }
-        });
-
-    }
-}
