+++ app/src/main/java/edu/wisc/ece/pinpoint/fragments/MapFragment.java
@@ -1,0 +1,21 @@
+package edu.wisc.ece.pinpoint.fragments;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import edu.wisc.ece.pinpoint.R;
+
+public class MapFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_map, container, false);
+    }
+}
