+++ app/src/main/res/animator/nav_default_enter_anim.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <!--Empty to disable animation-->
+</set>
