+++ app/src/main/res/layout/fragment_activity.xml
@@ -3,16 +3,14 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/black"
     tools:context=".fragments.ActivityFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginTop="148dp"
-        android:text="Activity Fragment!" />
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="Activity Fragment!"
+        android:textColor="@color/white"/>
 
 </RelativeLayout>
