+++ app/src/main/res/layout/fragment_profile_page.xml
@@ -1,15 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfilePageActivity">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraint_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+    tools:context=".fragments.ProfilePageFragment">
 
         <TextView
             android:id="@+id/profile_name"
@@ -145,60 +140,20 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/profile_bio" />
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-
     <com.google.android.material.tabs.TabLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_anchor="@+id/constraint_layout"
-        app:layout_anchorGravity="bottom"
-        android:layout_gravity="bottom"
+        app:layout_constraintTop_toBottomOf="@+id/profile_follow"
         android:id="@+id/tab_layout"
         app:tabRippleColor="@color/fui_transparent"
         />
 
-    <!-- Note that the page fragments are currently *wrap-content* height. This may need to be adjusted -->
     <androidx.viewpager2.widget.ViewPager2
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
         android:id="@+id/view_pager"
-        app:layout_anchor="@+id/tab_layout"
-        app:layout_anchorGravity="bottom"
-        android:layout_gravity="bottom"
+        app:layout_constraintTop_toBottomOf="@+id/tab_layout"
+        app:layout_constraintBottom_toBottomOf="parent"
         />
 
-
-    <com.google.android.material.bottomappbar.BottomAppBar
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        app:contentInsetStart="0dp"
-        android:id="@+id/bottomBar"
-        app:fabCradleMargin="10dp"
-        app:fabCradleRoundedCornerRadius="20dp"
-        >
-
-        <com.google.android.material.bottomnavigation.BottomNavigationView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:id="@+id/navBar"
-            app:menu="@menu/nav_bar"
-            android:background="@android:color/transparent"
-            android:outlineSpotShadowColor="@android:color/transparent"
-            app:itemRippleColor="@color/fui_transparent"
-            app:elevation="0dp"
-            />
-    </com.google.android.material.bottomappbar.BottomAppBar>
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_anchor="@id/navBar"
-        android:id="@+id/mapButton"
-        android:src="@drawable/ic_map"
-        app:fabCustomSize="56dp"
-        app:maxImageSize="32dp"
-        />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
