+++ app/src/main/res/menu/nav_bar.xml
@@ -9,7 +9,7 @@
         android:icon="@drawable/ic_leaderboard"
         android:title="Leaderboard" />
     <item
-        android:id="@+id/navbarEmpty"
+        android:id="@+id/navbarMap"
         android:title="" />
     <item
         android:id="@+id/navbarSearch"
